Fixed a hang: unsigned int -> NSUInteger

+++ License Utility/Classes/AQAppKit/OAGradientTableView.m
@@ -98,7 +98,7 @@ static const CGFunctionCallbacks linearFunctionCallbacks = {0, _linearColorBlend
     NSUInteger rowIndex = [selectedRowIndexes indexGreaterThanOrEqualToIndex:0];
 
     while (rowIndex != NSNotFound) {
-        unsigned int endOfCurrentRunRowIndex, newRowIndex = rowIndex;
+        NSUInteger endOfCurrentRunRowIndex, newRowIndex = rowIndex;
         do {
             endOfCurrentRunRowIndex = newRowIndex;
             newRowIndex = [selectedRowIndexes indexGreaterThanIndex:endOfCurrentRunRowIndex];
